Refactor date utility functions and update component imports for consistency
- Moved date-related utility functions to a new file `date-utils.ts` for better organization.
- Updated import paths in components to use the new utility file.
- Refactored `CalendarWidget`, `FocusMetrics`, `MyGoals`, and `Reminders` components to utilize the new date utility functions.
- Introduced new sub-components for better modularity in `CalendarEventList`, `CalendarWeekSelector`, `FocusMetricItem`, `GoalItem`, and `ReminderItem`.
