+++ src/components/my-goals.tsx
@@ -2,15 +2,13 @@
 
 import { Goal } from "lucide-react"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
-import { Progress } from "@/components/ui/progress"
-import {
-    Tooltip,
-    TooltipContent,
-    TooltipTrigger,
-} from "@/components/ui/tooltip"
+import { GoalItem } from "@/components/sub-components/goal-item"
 import { goals } from "@/data/dummy-data"
 
 export function MyGoals() {
+    const handleGoalClick = (goalId: number) => {
+        console.log(`Clicked goal: ${goalId}`)
+    }
 
     return (
         <Card className="rounded-xl border py-4 gap-2 shadow-none">
@@ -23,58 +21,11 @@ export function MyGoals() {
             <CardContent className="space-y-4 max-[560px]:space-y-3">
                 {goals.length > 0 ? (
                     goals.map((goal) => (
-                        <Tooltip key={goal.id} delayDuration={500}>
-                            <TooltipTrigger asChild>
-                                <div
-                                    className="p-2 pl-0 rounded-lg transition-colors cursor-pointer"
-                                >
-                                    {/* Desktop layout (≥560px) */}
-                                    <div className="hidden min-[560px]:flex items-start gap-4">
-                                        <div className="flex-1">
-                                            <h4 className="font-[580] text-md leading-tight">{goal.name}</h4>
-                                            <p className="text-[0.88rem] text-gray-light">
-                                                {goal.project} • My Projects
-                                            </p>
-                                        </div>
-
-                                        <div className="flex-1 max-w-[5rem] mt-2">
-                                            <Progress
-                                                value={goal.progress}
-                                                className="h-2 ml-2"
-                                            />
-                                        </div>
-                                        <span className="text-[0.94rem] font-[580] text-foreground min-w-[2.5rem] text-right">
-                                            {goal.progress}%
-                                        </span>
-                                    </div>
-
-                                    {/* Mobile layout (<560px) */}
-                                    <div className="block min-[560px]:hidden">
-                                        <div className="flex items-start justify-between gap-3 mb-2">
-                                            <div className="flex-1 min-w-0">
-                                                <h4 className="font-[580] text-sm leading-tight truncate">{goal.name}</h4>
-                                                <p className="text-[0.81rem] text-gray-medium mt-0.5 truncate">
-                                                    {goal.project} • My Projects
-                                                </p>
-                                            </div>
-                                            <span className="text-[0.94rem] font-[580] text-foreground ml-2 flex-shrink-0">
-                                                {goal.progress}%
-                                            </span>
-                                        </div>
-
-                                        <div className="w-full">
-                                            <Progress
-                                                value={goal.progress}
-                                                className="h-2"
-                                            />
-                                        </div>
-                                    </div>
-                                </div>
-                            </TooltipTrigger>
-                            <TooltipContent side="top" sideOffset={8}>
-                                <p className="text-sm font-medium leading-relaxed">{goal.name}</p>
-                            </TooltipContent>
-                        </Tooltip>
+                        <GoalItem
+                            key={goal.id}
+                            goal={goal}
+                            onClick={handleGoalClick}
+                        />
                     ))
                 ) : (
                     <div className="text-center py-8 text-muted-foreground">
