+++ src/components/reminders.tsx
@@ -2,16 +2,14 @@
 
 import { useState } from "react"
 import {
-    Bell,
     ChevronDown,
     ChevronUp,
-    Trash2,
     Timer,
-    Check
 } from "lucide-react"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Collapsible, CollapsibleContent, CollapsibleTrigger } from "@/components/ui/collapsible"
+import { ReminderItem } from "@/components/sub-components/reminder-item"
 import { reminders } from "@/data/dummy-data"
 
 export function Reminders() {
@@ -23,6 +21,10 @@ export function Reminders() {
 
     const handleReminderDelete = (reminderId: number) => {
         console.log(`Delete reminder: ${reminderId}`)
+    }
+
+    const handleReminderComplete = (reminderId: number) => {
+        console.log(`Complete reminder: ${reminderId}`)
     }
 
 
@@ -61,44 +63,13 @@ export function Reminders() {
                         </CollapsibleTrigger>
                         <CollapsibleContent className="space-y-3 mt-3">
                             {todayReminders.map((reminder) => (
-                                <div
+                                <ReminderItem
                                     key={reminder.id}
-                                    className="flex items-start justify-between gap-3 p-3 pl-0 transition-colors border-b-2 border-[#f2f3f7] cursor-pointer"
-                                >
-                                    <div className="flex items-start gap-3 flex-1">
-                                        <div className="flex-1 min-w-0">
-                                            <p className="text-[0.94rem] text-[#333333] leading-relaxed font-[550]">
-                                                {reminder.text}
-                                            </p>
-                                        </div>
-                                    </div>
-                                    <div className="flex items-center gap-1 sm:gap-2 flex-shrink-0">
-                                        <Button
-                                            variant="ghost"
-                                            size="sm"
-                                            className="size-9 p-0 hover:bg-blue-50 hover:text-blue-600"
-                                            onClick={() => handleReminderNotify(reminder.id)}
-                                        >
-                                            <Bell className="size-4" />
-                                        </Button>
-                                        <Button
-                                            variant="ghost"
-                                            size="sm"
-                                            className="size-9 p-0 hover:bg-red-50 hover:text-red-600"
-                                            onClick={() => handleReminderDelete(reminder.id)}
-                                        >
-                                            <Trash2 className="size-4" />
-                                        </Button>
-                                        <Button
-                                            variant="ghost"
-                                            size="sm"
-                                            className="size-6 p-0 bg-[#30d5d1] text-white rounded-full hover:bg-[#30d5d1]/80 hover:text-white"
-                                            onClick={() => handleReminderDelete(reminder.id)}
-                                        >
-                                            <Check />
-                                        </Button>
-                                    </div>
-                                </div>
+                                    reminder={reminder}
+                                    onNotify={handleReminderNotify}
+                                    onDelete={handleReminderDelete}
+                                    onComplete={handleReminderComplete}
+                                />
                             ))}
                         </CollapsibleContent>
                     </Collapsible>
