--- src/utils/get-the-c-date.tsx
@@ -1,9 +1,0 @@
-
-
-export const getTheCurrentDate = () => {
-    return new Date().toLocaleDateString('en-US', {
-        weekday: 'short',
-        month: 'long',
-        day: 'numeric'
-    })
-}
